requirements.txt
Pillow==10.0.1
Pillow==10.1.0
Requests==2.31.0
streamlit==1.28.2
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
